Echoes the contents of a file
This program will echo the contents of a file, whose path is given as the only
argument to the executable. The contents of the file are assumed to be stored in
a matrix-like format. This version will store the data in a matrix allocated as
a one dimensional block of memory before outputting it.