Makes a simple prediction
This version uses the minimum distance viewer to make a simple prediction for
movie 5 for a user. Instead of iterating the distance array to find the minimum
distance, the distances are stored in a struct, along with the viewer id to
which they are associated, which is then sorted. The minimum is then selected as
the first element in this struct array.